// generated/addon/angular/src/app/api-models/cieb-trans-accum-ident.model.ts
/* Copyright (c) 2021 . All Rights Reserved. */

export class CiebTransAccumIdent {
  seqAccumId: number;
  seqTransId: number;
  accumIdName: string;
  accumAdjType: string;
  systemType: string;
  trgtAccumAdjType: string;
  sysName: string;
  sysInitAdjName: string;
  sysAdjAmt: number;
  sysInitBalName: string;
  sysCombinedBalAmt: number;
  sysPristineBalAmt: number;
  sysAccumPeriodDt: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  errorCd: string;
}